mdbx: Cursor.Get keeps MDBX_RESULT_TRUE as success-with-data, drop dead Txn.key
/code-review found a real regression: gating on r.err != success turned
SetLowerBound/SetUpperBound inexact matches (MDBX_RESULT_TRUE) into
(nil, nil, nil) — old code routed the code through operrno, which maps
RESULT_TRUE to nil, and then built the slices. Restore that and add a
regression test. Also removes the now-unreferenced Txn.key scratch
field and folds the Get allocs test into allocs_test.go.
